model migration profile

+++ app/Http/Controllers/Admin/PlanController.php
@@ -113,10 +113,14 @@ class PlanController extends Controller
      */
     public function destroy($url)
     {
-        $plan = $this->repository->where('url', $url)->first();
+        $plan = $this->repository->with('details')->where('url', $url)->first();
 
         if (!$plan) {
             return redirect()->back();
+        }
+
+        if ($plan->details->count() > 0) {
+            return redirect()->back()->with('error', 'Existem detalhes vinculados!');
         }
 
         $plan->delete();
